+++ src/common/Button/Button.tsx
@@ -1,0 +1,38 @@
+import { h } from '@stencil/core';
+import { ButtonSizeEnum, ButtonVariantEnum } from './button.types';
+import styles from './button.styles';
+
+interface ButtonPropsType {
+    class?: string;
+    dataTestId?: string;
+    disabled?: boolean;
+    size?: `${ButtonSizeEnum}`;
+    variant?: `${ButtonVariantEnum}`;
+    onClick?: (event: MouseEvent) => void;
+}
+
+export function Button({ class: className = '', dataTestId = '', disabled = false, size = 'large', variant = 'primary', onClick }: ButtonPropsType, children?: any) {
+    return (
+        <button
+            data-testid={dataTestId}
+            onClick={onClick}
+            disabled={disabled}
+            class={{
+                [styles.button]: true,
+                [styles.buttonDisabled]: disabled,
+                [size]: Boolean(size),
+                [styles.buttonLarge]: size === ButtonSizeEnum.large,
+                [styles.buttonSmall]: size === ButtonSizeEnum.small,
+                [variant]: Boolean(variant),
+                [styles.buttonPrimary]: variant === ButtonVariantEnum.primary,
+                [styles.buttonSecondary]: variant === ButtonVariantEnum.secondary,
+                [styles.buttonSecondarySmall]: variant === ButtonVariantEnum.secondary && size === ButtonSizeEnum.small,
+                [styles.buttonNeutral]: variant === ButtonVariantEnum.neutral,
+                [className]: Boolean(className),
+            }}
+        >
+            {children}
+        </button>
+    );
+
+}