+++ src/common/CopyButton/getCopyClickAction.ts
@@ -1,0 +1,26 @@
+import { copyToClipboard } from 'utils/copyToClipboard';
+
+interface CopyHandlerOptions {
+    onSuccessChange?: (isSuccess: boolean) => void;
+}
+
+export function getCopyClickAction({ onSuccessChange }: CopyHandlerOptions) {
+    let timeoutId: number | null = null;
+
+    return async (event: MouseEvent, text?: string) => {
+        const trimmedText = text ? text.trim() : text;
+        const success = await copyToClipboard(trimmedText);
+
+        event.preventDefault();
+        event.stopPropagation();
+
+        if (onSuccessChange) {
+            onSuccessChange(success);
+
+            window.clearTimeout(timeoutId);
+            if (success) {
+                timeoutId = window.setTimeout(() => onSuccessChange(false), 2000);
+            }
+        }
+    };
+}