+++ src/components/functional/sign-transactions-panel/helpers/handleAmountResize/handleAmountResize.ts
@@ -7,7 +7,7 @@ export const handleAmountResize = (element: HTMLElement | null) => {
   const getFontSize = (element: HTMLElement) => parseInt(getComputedStyle(element).getPropertyValue('font-size'));
 
   const firstChild = element.firstChild as HTMLElement;
-  const maxWidth = 270;
+  const maxWidth = 250;
   const sizes = {
     parent: Math.min(element.offsetWidth, maxWidth),
     firstChild: getFontSize(firstChild),